# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: الدمشقي ERP + e-commerce + POS + manufacturing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales-orders:
    get:
      operationId: listSalesOrders
      tags: [orders]
      parameters:
        - name: channel
          in: query
          required: false
          schema: { type: string, enum: [pos, online] }
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: dateFrom
          in: query
          required: false
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          required: false
          schema: { type: string, format: date }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SalesOrder" }
    post:
      operationId: createSalesOrder
      tags: [orders, pos]
      summary: Create POS sale (immediate fulfillment)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PosOrderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }
  /sales-orders/{id}:
    get:
      operationId: getSalesOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }
  /sales-orders/{id}/status:
    patch:
      operationId: updateSalesOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/OrderStatus" }
                note: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }
...
components:
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, preparing, ready, out_for_delivery, completed, cancelled]
    SalesOrder:
      type: object
      required: [id, orderNumber, channel, status, subtotalMinor, totalMinor, currency, items, createdAt]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        channel: { type: string, enum: [pos, online] }
        status: { $ref: "#/components/schemas/OrderStatus" }
        customerId: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        customerEmail: { type: string }
        deliveryAddress: { type: string }
        deliveryNotes: { type: string }
        paymentMethod: { type: string, enum: [cash, card, cod, online] }
        subtotalMinor: { type: integer }
        discountMinor: { type: integer }
        deliveryFeeMinor: { type: integer }
        totalMinor: { type: integer }
        currency: { type: string }
        cashierId: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/SalesOrderItem" }
        createdAt: { type: string, format: date-time }
    SalesOrderItem:
      type: object
      required: [id, productId, productNameAr, quantity, unitPriceMinor, lineTotalMinor]
      properties:
        id: { type: string }
        productId: { type: string }
        productNameAr: { type: string }
        quantity: { type: integer }
        unitPriceMinor: { type: integer }
        lineTotalMinor: { type: integer }
    PosOrderInput:
      type: object
      required: [items, paymentMethod]
      properties:
        customerName: { type: string }
        customerPhone: { type: string }
        paymentMethod: { type: string, enum: [cash, card] }
        discountMinor: { type: integer }
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string }
              quantity: { type: integer, minimum: 1 }
...